Move kerberos client building logic to its own file to make it accessible from tests.
