+++ .travis.yml
@@ -10,9 +10,9 @@ before_install:
 - "$HOME/bats/install.sh $HOME/bats/build"
 - export PATH="$PATH:$HOME/bats/build/bin"
 env:
-- HADOOP_DISTRO=cdh KERBEROS=true HADOOP_CONF_DIR="/tmp/kdc-home/"
+- HADOOP_DISTRO=cdh KERBEROS=true HADOOP_CONF_DIR="/tmp/kdc-home/" HADOOP_KRB_CONF=/tmp/kdc-home/krb5.conf HADOOP_KEYTAB=/tmp/kdc-home/client.keytab
 - HADOOP_DISTRO=cdh KERBEROS=false
-- HADOOP_DISTRO=hdp KERBEROS=true HADOOP_CONF_DIR="/tmp/kdc-home/"
+- HADOOP_DISTRO=hdp KERBEROS=true HADOOP_CONF_DIR="/tmp/kdc-home/" HADOOP_KRB_CONF=/tmp/kdc-home/krb5.conf HADOOP_KEYTAB=/tmp/kdc-home/client.keytab
 - HADOOP_DISTRO=hdp KERBEROS=false
 before_script:
 - export NN_PORT=9000
