+++ cmd/hdfs/main.go
@@ -4,23 +4,13 @@ import (
 	"fmt"
 	"log"
 	"os"
-	"os/user"
 	"strings"
 
 	"github.com/colinmarc/hdfs"
 	"github.com/pborman/getopt"
-	"gopkg.in/jcmturner/gokrb5.v3/client"
-	"gopkg.in/jcmturner/gokrb5.v3/config"
-	"gopkg.in/jcmturner/gokrb5.v3/credentials"
-	"gopkg.in/jcmturner/gokrb5.v3/keytab"
 )
 
 // TODO: cp, tree, test, trash
-
-const hdfsDefaultServiceName = "nn"
-const hdfsDefaultConfDir = "/etc/hadoop/conf"
-const krbDefaultCfgPath = "/etc/krb5.conf"
-const hadoopAuthCfgPath = "hadoop.security.authentication"
 
 var (
 	version string
@@ -52,7 +42,7 @@ the following environment variables may be used:
   - HADOOP_KRB_CONF     kerberos configuration file. Default: %s
   - HADOOP_CCACHE       credential cache to use. Defaults: to "/tmp/krb5cc_{user_uid}"
   - HADOOP_KEYTAB       if set, the specified keytab is used and the credential cache is ignored.
-`, os.Args[0], hdfsDefaultConfDir, krbDefaultCfgPath)
+`, os.Args[0], hdfs.HdfsDefaultConfDir, hdfs.KrbDefaultCfgPath)
 
 	lsOpts = getopt.New()
 	lsl    = lsOpts.Bool('l')
@@ -196,7 +186,7 @@ func getClient(namenodes string) (*hdfs.Client, error) {
 		return cachedClient, nil
 	}
 
-	hadoopCfg := loadHadoopConf()
+	hadoopCfg := hdfs.LoadHadoopConf(hdfs.GetConfDir())
 
 	options := hdfs.ClientOptions{}
 	if namenodes != "" {
@@ -205,8 +195,8 @@ func getClient(namenodes string) (*hdfs.Client, error) {
 		options.Addresses = getNameNodes(hadoopCfg)
 	}
 
-	options.KerberosClient = getKrbClientIfRequired(hadoopCfg)
-	options.ServicePrincipalName = getServiceName()
+	options.KerberosClient = hdfs.GetKrbClientIfRequired(hadoopCfg)
+	options.ServicePrincipalName = hdfs.GetServiceName()
 
 	c, err := hdfs.NewClient(options)
 	if err != nil {
@@ -216,27 +206,6 @@ func getClient(namenodes string) (*hdfs.Client, error) {
 	cachedClient = c
 
 	return cachedClient, nil
-}
-
-// getServiceName returns 'nn' unless the HADOOP_SNAME environment variable is set
-func getServiceName() string {
-	if sn := os.Getenv("HADOOP_SNAME"); sn != "" {
-		return sn
-	}
-	return hdfsDefaultServiceName
-}
-
-// getConfDir returns the content of HADOOP_CONF_DIR or a default path to the conf dir.
-func getConfDir() string {
-	if cd := os.Getenv("HADOOP_CONF_DIR"); cd != "" {
-		return cd
-	}
-	return hdfsDefaultConfDir
-}
-
-// loadHadoopConf attempts to load the hadop configuration from a specified or default path.
-func loadHadoopConf() hdfs.HadoopConf {
-	return hdfs.LoadHadoopConf(getConfDir())
 }
 
 // getNameNodes checks the HADOOP_NAMENODE or the passed configuration for the namenode servers
@@ -253,106 +222,4 @@ func getNameNodes(conf hdfs.HadoopConf) []string {
 	}
 
 	return nn
-}
-
-// getKrbClientIfRequired returns a client if the hadoop configuration or the environment variables suggest one is required
-func getKrbClientIfRequired(conf hdfs.HadoopConf) *client.Client {
-	// Check if the kerberos config path has been overriden
-	var krb5Cfg = os.Getenv("HADOOP_KRB_CONF")
-
-	if krb5Cfg == "" {
-		krb5Cfg = krbDefaultCfgPath
-	}
-
-	// Now check if the credential cache or the keytab have been manually specified
-	keytabPath := os.Getenv("HADOOP_KEYTAB")
-
-	if keytabPath != "" {
-		return getKrbClientWithKeytab(krb5Cfg, keytabPath)
-	}
-
-	credCachePath := os.Getenv("HADOOP_CCACHE")
-	if credCachePath != "" {
-		return getKrbClientWithCredCache(krb5Cfg, credCachePath)
-	}
-
-	// At this point none of the required env vars have been required
-	// and we need to check the configuration to know if kerberos is enabled
-	val, found := conf[hadoopAuthCfgPath]
-	if !found || "kerberos" != strings.ToLower(val) {
-		return nil
-	}
-
-	// Kerberos is enabled. fall back to the default credential cache location.
-	// TODO: read the kerberos config to determine where the cred cache is located?
-	return getKrbClientWithCredCache(krb5Cfg, getDefaultCredCachePath())
-}
-
-// returns "/tmp/krb5cc_$(id -u $(whoami))"
-func getDefaultCredCachePath() string {
-	u, e := user.Current()
-	if e != nil {
-		log.Panic(e)
-	}
-	return "/tmp/krb5cc_" + u.Uid
-}
-
-func getKrbClientWithCredCache(configPath string, cachePath string) *client.Client {
-	cfg, cfgE := config.Load(configPath)
-
-	if cfgE != nil {
-		log.Panic(cfgE)
-	}
-
-	cc, cce := credentials.LoadCCache(cachePath)
-
-	if cce != nil {
-		log.Panic(cce)
-	}
-
-	cl, clE := client.NewClientFromCCache(cc)
-	if clE != nil {
-		log.Panic(clE)
-	}
-
-	cl.WithConfig(cfg)
-	// TODO Config flag or whatever for people not using AD
-	cl.GoKrb5Conf.DisablePAFXFast = true
-
-	return &cl
-
-}
-
-func getKrbClientWithKeytab(configPath string, keytabPath string) *client.Client {
-
-	cfg, cfgE := config.Load(configPath)
-
-	if cfgE != nil {
-		log.Panic(cfgE)
-	}
-
-	kt, ktE := keytab.Load(keytabPath)
-
-	if ktE != nil {
-		log.Panic(ktE)
-	}
-
-	entries := kt.Entries
-
-	if len(entries) == 0 {
-		log.Fatalf("no entries found in keytab %s" + keytabPath)
-	}
-
-	// Fetch the principal of the first entry
-	principal := entries[0].Principal
-
-	cl := client.NewClientWithKeytab(strings.Join(principal.Components, "/"), principal.Realm, kt)
-	cl.WithConfig(cfg)
-
-	// TODO Config flag or whatever for people not using AD
-	cl.GoKrb5Conf.DisablePAFXFast = true
-	if loginE := cl.Login(); loginE != nil {
-		log.Panic(loginE)
-	}
-	return &cl
 }
