+++ setup_kdc.sh
@@ -25,12 +25,16 @@ sleep 3
 
 # Generate the user and the service keytabs.
 echo "Adding user principal and generating keytab..."
-echo "addprinc -pw password bob@EXAMPLE.COM\nktadd -k bob.keytab bob@EXAMPLE.COM\nexit" | ./kadmin.sh $SRV_CONF_DIR -k $KEYTAB_DIR/admin.keytab
+echo "addprinc -pw password $USER@EXAMPLE.COM\nktadd -k client.keytab $USER@EXAMPLE.COM\nexit" | ./kadmin.sh $SRV_CONF_DIR -k $KEYTAB_DIR/admin.keytab
 
 echo "Adding service principal and generating keytab..."
 echo "addprinc -randkey nn/localhost@EXAMPLE.COM\nktadd -k nn.keytab nn/localhost@EXAMPLE.COM\nexit" | ./kadmin.sh $SRV_CONF_DIR -k $KEYTAB_DIR/admin.keytab
-echo "Moving the keytabs to /tmp/kdc-home/"
-mv bob.keytab /tmp/kdc-home/
-mv nn.keytab /tmp/kdc-home/
+echo "Moving the keytabs to $KDC_HOME"
+mv client.keytab $KDC_HOME
+mv nn.keytab $KDC_HOME
 cp krb5.conf $KDC_HOME
 
+# Initialize a credential cache.
+./kinit.sh -conf $KDC_HOME -k -t $KDC_HOME/client.keytab  $USER@EXAMPLE.COM
+# Move the credential cache to where the native client expects it
+cp /tmp/krb5cc_hdfs_test /tmp/krb5cc_$(id -u $(whoami))
