+++ test/kdc/krb5.conf
@@ -5,10 +5,10 @@
   ticket_lifetime = 24h
   dns_lookup_realm = false
   dns_lookup_kdc = false
-  default_ccache_name = /tmp/krb5cc_%{uid}
+  # explicitly NOT using the {uid} substitution:
+  # kerby's kinit apprently does not understand it
+  default_ccache_name = /tmp/krb5cc_hdfs_test
   udp_preference_limit = 1
-#  default_tgs_enctypes = aes
-#  default_tkt_enctypes = aes
 
 
 [realms]
